feat(odyssey-react-mui): create hook for reusable autocompletelogic
